fix: update interfaces
Updates to latest code from https://github.com/libp2p/js-libp2p-interfaces/pull/180
