+++ package.json
@@ -3,13 +3,13 @@
   "version": "1.0.1",
   "description": "JavaScript implementation of https://github.com/libp2p/mplex",
   "license": "Apache-2.0 OR MIT",
-  "homepage": "https://github.com/libp2p/js-libp2p-mplex#readme",
+  "homepage": "https://github.com/libp2p/js-libp2p-interfaces/tree/master/packages/libp2p-mplex#readme",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/libp2p/js-libp2p-mplex.git"
+    "url": "git+https://github.com/libp2p/js-libp2p-interfaces.git"
   },
   "bugs": {
-    "url": "https://github.com/libp2p/js-libp2p-mplex/issues"
+    "url": "https://github.com/libp2p/js-libp2p-interfaces/issues"
   },
   "keywords": [
     "IPFS",
@@ -127,10 +127,10 @@
   },
   "scripts": {
     "lint": "aegir lint",
-    "dep-check": "aegir dep-check",
+    "dep-check": "aegir dep-check dist/src/**/*.js dist/test/**/*.js",
     "build": "tsc",
     "pretest": "npm run build",
-    "test": "aegir test -f ./dist/test",
+    "test": "aegir test -f ./dist/test/*.js -f ./dist/test/**/*.js",
     "test:chrome": "npm run test -- -t browser --cov",
     "test:chrome-webworker": "npm run test -- -t webworker",
     "test:firefox": "npm run test -- -t browser -- --browser firefox",
@@ -140,24 +140,24 @@
     "release": "semantic-release"
   },
   "dependencies": {
-    "@libp2p/logger": "^1.0.3",
-    "@libp2p/tracked-map": "^1.0.1",
+    "@libp2p/logger": "^1.1.2",
+    "@libp2p/tracked-map": "^1.0.4",
     "abortable-iterator": "^4.0.2",
     "any-signal": "^3.0.0",
     "err-code": "^3.0.1",
     "it-pipe": "^2.0.3",
     "it-pushable": "^2.0.1",
     "it-stream-types": "^1.0.4",
-    "uint8arraylist": "^1.2.0",
+    "uint8arraylist": "^1.4.0",
     "varint": "^6.0.0"
   },
   "devDependencies": {
-    "@libp2p/interface-compliance-tests": "^1.0.7",
-    "@libp2p/interfaces": "^1.1.1",
+    "@libp2p/interface-compliance-tests": "^1.1.16",
+    "@libp2p/interfaces": "^1.3.14",
     "@types/varint": "^6.0.0",
     "aegir": "^36.1.3",
-    "cborg": "^1.2.1",
-    "iso-random-stream": "^2.0.0",
+    "cborg": "^1.8.1",
+    "iso-random-stream": "^2.0.2",
     "it-all": "^1.0.6",
     "it-drain": "^1.0.5",
     "it-foreach": "^0.1.1",
